login-register-backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=3.15
pymysql>=1.0.4
python-dotenv>=1.0.0
